# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Solana Meme Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BotStatus {
  running: boolean;
  /** @nullable */
  startedAt: string | null;
  /** @nullable */
  uptime: number | null;
  tokensScanned: number;
  /** @nullable */
  lastScanAt: string | null;
}

export interface BotConfig {
  id: number;
  /** @nullable */
  walletAddress: string | null;
  buyMarketCapUsd: number;
  sellMarketCapUsd: number;
  maxPositions: number;
  slippageBps: number;
  updatedAt: string;
}

export interface BotConfigInput {
  walletAddress?: string;
  buyMarketCapUsd?: number;
  sellMarketCapUsd?: number;
  maxPositions?: number;
  slippageBps?: number;
}

export type PositionStatus = typeof PositionStatus[keyof typeof PositionStatus];


export const PositionStatus = {
  open: 'open',
  closed: 'closed',
  pending_sell: 'pending_sell',
} as const;

export interface Position {
  id: number;
  tokenAddress: string;
  tokenSymbol: string;
  tokenName: string;
  entryMarketCapUsd: number;
  /** @nullable */
  currentMarketCapUsd: number | null;
  entryPriceUsd: number;
  /** @nullable */
  currentPriceUsd: number | null;
  amountTokens: number;
  amountUsd: number;
  /** @nullable */
  unrealizedPnlUsd: number | null;
  /** @nullable */
  unrealizedPnlPct: number | null;
  status: PositionStatus;
  openedAt: string;
  /** @nullable */
  closedAt: string | null;
}

export type TradeType = typeof TradeType[keyof typeof TradeType];


export const TradeType = {
  buy: 'buy',
  sell: 'sell',
} as const;

export type TradeStatus = typeof TradeStatus[keyof typeof TradeStatus];


export const TradeStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  failed: 'failed',
} as const;

export interface Trade {
  id: number;
  tokenAddress: string;
  tokenSymbol: string;
  tokenName: string;
  type: TradeType;
  amountUsd: number;
  amountTokens: number;
  priceUsd: number;
  marketCapUsd: number;
  /** @nullable */
  txSignature: string | null;
  status: TradeStatus;
  createdAt: string;
  /** @nullable */
  realizedPnlUsd: number | null;
}

export interface WatchedToken {
  address: string;
  symbol: string;
  name: string;
  marketCapUsd: number;
  priceUsd: number;
  volume24h: number;
  priceChange24h: number;
  liquidity: number;
  /** @nullable */
  pairAddress: string | null;
  /** @nullable */
  dexId: string | null;
  firstDiscoveredAt: string;
}

export interface DashboardSummary {
  totalTrades: number;
  openPositions: number;
  totalRealizedPnlUsd: number;
  totalUnrealizedPnlUsd: number;
  winRate: number;
  tokensWatched: number;
  botRunning: boolean;
  /** @nullable */
  walletAddress: string | null;
}

export type ActivityEntryType = typeof ActivityEntryType[keyof typeof ActivityEntryType];


export const ActivityEntryType = {
  bot_started: 'bot_started',
  bot_stopped: 'bot_stopped',
  token_discovered: 'token_discovered',
  buy_executed: 'buy_executed',
  sell_executed: 'sell_executed',
  buy_failed: 'buy_failed',
  sell_failed: 'sell_failed',
  scan_completed: 'scan_completed',
} as const;

export interface ActivityEntry {
  id: number;
  type: ActivityEntryType;
  message: string;
  /** @nullable */
  tokenSymbol: string | null;
  /** @nullable */
  tokenAddress: string | null;
  createdAt: string;
}

export type ListTradesParams = {
limit?: number;
offset?: number;
};