# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Solana Meme Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /positions/{id}:
    get:
      operationId: getPosition
      tags: [positions]
      summary: Get a specific position
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Position details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
...
components:
  schemas:
...
    Position:
      type: object
      properties:
        id:
          type: integer
        tokenAddress:
          type: string
        tokenSymbol:
          type: string
        tokenName:
          type: string
        entryMarketCapUsd:
          type: number
        currentMarketCapUsd:
          type: ["number", "null"]
        entryPriceUsd:
          type: number
        currentPriceUsd:
          type: ["number", "null"]
        amountTokens:
          type: number
        amountUsd:
          type: number
        unrealizedPnlUsd:
          type: ["number", "null"]
        unrealizedPnlPct:
          type: ["number", "null"]
        status:
          type: string
          enum: [open, closed, pending_sell]
        openedAt:
          type: string
        closedAt:
          type: ["string", "null"]
      required:
        - id
        - tokenAddress
        - tokenSymbol
        - tokenName
        - entryMarketCapUsd
        - currentMarketCapUsd
        - entryPriceUsd
        - currentPriceUsd
        - amountTokens
        - amountUsd
        - unrealizedPnlUsd
        - unrealizedPnlPct
        - status
        - openedAt
        - closedAt
...