# lib/api-zod/src/generated/types/position.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Solana Meme Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { PositionStatus } from './positionStatus';

export interface Position {
  id: number;
  tokenAddress: string;
  tokenSymbol: string;
  tokenName: string;
  entryMarketCapUsd: number;
  /** @nullable */
  currentMarketCapUsd: number | null;
  entryPriceUsd: number;
  /** @nullable */
  currentPriceUsd: number | null;
  amountTokens: number;
  amountUsd: number;
  /** @nullable */
  unrealizedPnlUsd: number | null;
  /** @nullable */
  unrealizedPnlPct: number | null;
  status: PositionStatus;
  openedAt: string;
  /** @nullable */
  closedAt: string | null;
}